Introduced abstract Directive and moved all behaviour from Allow & Deny
